fs_scans: dual backend (SQLite + CNPG PostgreSQL) via consolidation (#75)
* committing plan

* fs_scans: add backend abstraction (sqlite default + postgres)

Introduce a dual-backend configuration layer for fs_scans, mirroring the
existing job_history pattern, so the same ORM/query code can target either
local SQLite .db files (the default, unchanged) or a shared PostgreSQL
(CNPG) server.

Why: we want to additionally publish the weekly scan data to a networked
database visible on a webserver, without disturbing the fast SQLite
generation pipeline or local CLI access.

- New fs_scans/core/config.py (FsScanConfig): FS_SCAN_DB_BACKEND selector,
  FS_SCAN_PG_* connection settings, PG_DB_NAME (database == filesystem),
  pg_schema_name() (schema == collection), validate_postgres().
- database.py: get_engine() branches on backend; postgres pins search_path
  to the collection schema so existing bare-table SQL resolves unchanged.
  get_db_path() now raises for non-sqlite; add get_db_url(), db_available(),
  _ensure_pg_database(); init_db()/drop_tables()/get_session() gain a
  schema kwarg. SQLite behavior is unchanged.
- .env.example: document the FS_SCAN_* block.


* fs_scans: make query/read path portable to PostgreSQL

The query layer had two SQLite-only constructs that would fail on the
PostgreSQL read path served to the webserver:

- query_builder name matching used the SQLite-only GLOB operator. Thread the
  bind dialect into DirectoryQueryBuilder and emit an anchored regex (`~`) for
  case-sensitive matches and ILIKE for case-insensitive matches on postgresql,
  keeping GLOB/LIKE on sqlite. New glob_to_posix_regex() translates the
  `*`/`?` wildcard subset and escapes regex metacharacters.
- histogram_common probed table existence via `sqlite_master`; replace with
  dialect-agnostic SQLAlchemy inspect().has_table().

Also guard the import-time SQLite PRAGMAs (file_handling) to no-op on
non-sqlite binds. SQLite behavior is unchanged; generation path untouched.


* fs_scans: backend-aware database discovery + CLI

get_all_filesystems(), --show-config, and the single-collection existence
checks assumed SQLite files and would break (or raise from get_db_path) under
the postgres backend.

- database.py: add list_pg_schemas() (collections == schemas, excluding
  system and transient *_staging/*_old schemas), filesystem_available()
  (backend-aware existence), and describe_databases() (name/location/size for
  display).
- query_engine.get_all_filesystems(): glob .db files for sqlite, delegate to
  list_pg_schemas() for postgres.
- common.render_show_config(): shared, backend-aware --show-config renderer
  (SQLite data dir + sizes, or PG server + schemas); query and analyze
  commands now use it and filesystem_available()/get_db_url() instead of
  get_db_path().exists().


* fs_scans: deterministic query ordering via dir_id tiebreaker

The directory query ORDER BY had no tiebreaker, so rows with equal sort keys
(e.g. equal total_size_r) came back in arbitrary, engine-dependent order. This
was latent on SQLite alone and became visible comparing SQLite vs PostgreSQL:
identical sort-key sequences but different paths, and LIMIT cutting through a
tie group kept different equal-keyed rows on each engine.

Append the unique, verbatim-copied dir_id as a final ORDER BY key so results
are fully deterministic and identical across backends.


* fs_scans: add SQLite -> PostgreSQL consolidation step

Add `fs-scans consolidate`, which loads a finished per-collection SQLite .db
into a PostgreSQL schema and atomically swaps it into place — so the weekly
scan data can additionally be served from the networked CNPG database without
touching the fast SQLite generation pipeline.

Per collection (database = filesystem `campaign`, schema = collection):
- (re)create a <collection>_staging schema + tables; drop FKs so the bulk
  load is not validated row-by-row.
- stream every table from the read-only SQLite source via PostgreSQL COPY
  (bounded memory). DateTime columns are encoded specially: SQLite's dynamic
  typing allows an integer-epoch sentinel (0 = "no atime") in a DateTime
  column, which is mapped to a real timestamp so it keeps sorting before real
  timestamps and Postgres TIMESTAMP accepts it.
- deferred indexing: re-add (validated) FKs, build the same secondary index
  set as the SQLite import, ANALYZE.
- grant SELECT to the read-only web role (FS_SCAN_PG_READONLY_ROLE, optional).
- atomic swap via ALTER SCHEMA RENAME, dropping the previous generation to
  reclaim disk (--keep-old to retain, --no-swap to stage only, --no-fk to skip
  FK validation).

Per-collection failures are isolated so one bad .db cannot block the others.
Includes pure-Python encoder/stream tests (always run) and an opt-in live
round-trip + swap test (FS_SCAN_RUN_PG_TESTS=1).


* fs_scans: weekly consolidation PBS job + docs

Wire the SQLite→PostgreSQL consolidation into the weekly pipeline and document
the dual-backend architecture.

- PBS/consolidate.pbs: runs `fs-scans-consolidate --all` against CNPG after the
  collect_results rsync. Modest local resources (work is server-side); exports
  FS_SCAN_PG_* from .env so it works without an editable install.
- PBS/submit_all.sh: capture the collect_results job id and submit consolidate
  with depend=afterok. Gated behind FS_SCAN_ENABLE_CONSOLIDATE=1 so it stays
  OFF until the per-collection performance pass and CNPG disk/role review sign
  off.
- consolidator: raise maintenance_work_mem (FS_SCAN_PG_MAINTENANCE_WORK_MEM,
  default 1GB) for the post-load index builds.
- CLAUDE.md / fs_scans/README.md: document backends, schema-per-collection
  layout, the consolidate command, and the atomic swap.


* fs_scans: widen UID/GID columns to BigInteger for PostgreSQL

Filesystem scans contain unmapped/squashed owners with 32-bit *unsigned* UIDs
(e.g. nobody = 4294967294 = 2^32-2). SQLite's 64-bit INTEGER stored these fine,
but PostgreSQL `integer` is signed 32-bit (max 2,147,483,647), so consolidating
such a collection failed: "value 4294967294 is out of range for type integer".

Widen every UID/GID-bearing column to BigInteger: directory_stats.owner_uid/
owner_gid, the owner_summary/group_summary PKs, user_info.uid, group_info.gid,
and both histogram owner_uid PKs. dir_id/parent_id/depth/bucket_index stay
32-bit (row ids / small ints, never UIDs). SQLite is unaffected (dynamic typing);
verified by re-consolidating `collections` (2,620 rows at uid 4294967294).


* fs_scans: enable weekly CNPG consolidation in the cron driver

The SQLite→PostgreSQL consolidation is validated (all 13 collections, ~50 GB in
PG ≈ 1.14×, sub-second atomic swaps, full SQLite↔PostgreSQL parity), so turn it
on for the automated weekly run by exporting FS_SCAN_ENABLE_CONSOLIDATE=1 in
cron_driver.sh. Manual submit_all.sh invocations stay opt-in.


* fs_scans: fix consolidate.pbs repo resolution under PBS

The weekly consolidate job failed at startup with "not a git repository"
and a bogus /etc/config_env.sh path. Under PBS the script executes from a
spool copy (/var/spool/pbs/mom_priv/jobs/...), so deriving TOP_DIR via
`git -C ${SCRIPT_DIR}` ran git inside the spool dir — not a work tree —
leaving TOP_DIR empty.

Resolve the repo from the job's working directory instead (PBS starts in
the submission dir; falls back to CWD for manual runs), matching the
proven pattern in fs_scan.pbs and rsync_results.pbs.


---------

+++ fs_scans/core/models.py
@@ -79,9 +79,11 @@ class DirectoryStats(Base):
     dir_count_r = Column(BigInteger, nullable=False, default=0)
     max_atime_r = Column(DateTime, nullable=True)
 
-    # Owner/group tracking: -1=no files yet, NULL=multiple, else=single UID/GID
-    owner_uid = Column(Integer, nullable=True, default=-1)
-    owner_gid = Column(Integer, nullable=True, default=-1)
+    # Owner/group tracking: -1=no files yet, NULL=multiple, else=single UID/GID.
+    # BigInteger: 32-bit *unsigned* UIDs/GIDs (e.g. nobody=4294967294) exceed
+    # PostgreSQL's 32-bit signed `integer` range.
+    owner_uid = Column(BigInteger, nullable=True, default=-1)
+    owner_gid = Column(BigInteger, nullable=True, default=-1)
 
     # Relationship
     directory = relationship("Directory", back_populates="stats")
@@ -140,7 +142,7 @@ class OwnerSummary(Base):
 
     __tablename__ = "owner_summary"
 
-    owner_uid = Column(Integer, primary_key=True)
+    owner_uid = Column(BigInteger, primary_key=True)
     total_size = Column(BigInteger, default=0)
     total_files = Column(BigInteger, default=0)
     directory_count = Column(Integer, default=0)
@@ -161,7 +163,7 @@ class UserInfo(Base):
 
     __tablename__ = "user_info"
 
-    uid = Column(Integer, primary_key=True)
+    uid = Column(BigInteger, primary_key=True)
     username = Column(Text)
     full_name = Column(Text)  # GECOS field
 
@@ -178,7 +180,7 @@ class GroupInfo(Base):
 
     __tablename__ = "group_info"
 
-    gid = Column(Integer, primary_key=True)
+    gid = Column(BigInteger, primary_key=True)
     groupname = Column(Text)
 
     def __repr__(self):
@@ -194,7 +196,7 @@ class GroupSummary(Base):
 
     __tablename__ = "group_summary"
 
-    owner_gid = Column(Integer, primary_key=True)
+    owner_gid = Column(BigInteger, primary_key=True)
     total_size = Column(BigInteger, default=0)
     total_files = Column(BigInteger, default=0)
     directory_count = Column(Integer, default=0)
@@ -215,7 +217,7 @@ class AccessHistogram(Base):
 
     __tablename__ = "access_histogram"
 
-    owner_uid = Column(Integer, primary_key=True)
+    owner_uid = Column(BigInteger, primary_key=True)
     bucket_index = Column(Integer, primary_key=True)  # 0-9 (maps to ATIME_BUCKETS)
 
     file_count = Column(BigInteger, default=0)
@@ -279,7 +281,7 @@ class SizeHistogram(Base):
 
     __tablename__ = "size_histogram"
 
-    owner_uid = Column(Integer, primary_key=True)
+    owner_uid = Column(BigInteger, primary_key=True)
     bucket_index = Column(Integer, primary_key=True)  # 0-9 (maps to SIZE_BUCKETS)
 
     file_count = Column(BigInteger, default=0)
